Server bug fix. Arbiter bug fix. Still WIP
